backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
python-dotenv
google-genai

pydantic
pydantic-settings
urllib3
